Added a directory check

+++ main.js
@@ -13,20 +13,21 @@ program.parse(process.argv);
 const options = program.opts();
 const host = options.host;
 const port = options.port;
-const cache = options.cache;
+let cache = options.cache;
+cache = "./cache2";
 
-if (!host) {
-  console.error("Please, specify server address (host)");
-  process.exit(1);
-}
-if (!port) {
-  console.error("Please, specify server port number");
-  process.exit(1);
-}
-if (!cache) {
-  console.error("Please, specify the path to the directory with cached files");
-  process.exit(1);
-}
+// if (!host) {
+//   console.error("Please, specify server address (host)");
+//   process.exit(1);
+// }
+// if (!port) {
+//   console.error("Please, specify server port number");
+//   process.exit(1);
+// }
+// if (!cache) {
+//   console.error("Please, specify the path to the directory with cached files");
+//   process.exit(1);
+// }
 
 const server = http.createServer((req, res) => {
   const url = req.url;
@@ -36,8 +37,13 @@ const server = http.createServer((req, res) => {
     return;
   }
 
+  // якщо такої теки не існує, то її треба створити
+  if (!fs.existsSync(cache)) {
+    fs.mkdirSync(cache);
+  }
+
   if (req.method === "GET") {
-    // Ігноруємо запит до favicon.ico
+    // ігноруємо запит до favicon.ico
     if (url === "/favicon.ico") {
       res.writeHead(204); // 204 No Content
       res.end();
@@ -57,7 +63,7 @@ const server = http.createServer((req, res) => {
         .get(`https://http.cat${url}`)
         .then((response) => {
           const data = response.body;
-          // Зберігаю нове зображення у кеш
+          // зберігаю нове зображення у кеш
           fs.promises.writeFile(filePath, data).then(() => {
             res.setHeader("Content-Type", "image/jpeg");
             res.writeHead(200);
@@ -78,11 +84,12 @@ const server = http.createServer((req, res) => {
     });
 
     req.on("end", () => {
-      const buffer = Buffer.concat(body); // Об'єднуємо всі частини в один буфер
+      // об'єднуємо всі частини в один буфер
+      const buffer = Buffer.concat(body);
 
       const filePath = path.join(cache, `${url}.jpeg`);
 
-      // Записуємо файл
+      // записуємо файл
       fs.writeFile(filePath, buffer, (err) => {
         if (!err) {
           res.writeHead(201);
@@ -94,21 +101,34 @@ const server = http.createServer((req, res) => {
       });
     });
   } else if (req.method === "DELETE") {
-    fs.unlink(filePath, (err) => {
-      if (!err) {
-        res.writeHead(200);
-        res.end("File deleted successfully");
+    const filePath = path.join(cache, `${url}.jpeg`);
+    if (fs.existsSync(filePath)) {
+      fs.unlink(filePath, (err) => {
+        if (!err) {
+          res.writeHead(200);
+          res.end("File deleted successfully");
+          return;
+        }
+        res.writeHead(500);
+        res.end("Server error");
         return;
-      }
+      });
+    } else {
+      res.setHeader("Content-Type", "text/html");
       res.writeHead(404);
-      res.end("File not found");
-    });
+      res.end("<h1>404 Not found</h1>");
+      return;
+    }
   } else {
     res.writeHead(405);
     res.end("Method not allowed");
   }
 });
 
-server.listen(port, host, () => {
-  console.log(`Сервер запущений на http://${host}:${port}`);
+// server.listen(port, host, () => {
+//   console.log(`Сервер запущений на http://${host}:${port}`);
+// });
+
+server.listen("3000", "127.0.0.1", () => {
+  console.log(`Сервер запущений на http://${options.host}:${options.port}`);
 });
